native: unsigned long -> u32 in 14 game headers (LP64 vtable/signature match)
The decomp convention is 'unsigned long' == 32-bit (GC ILP32). On LP64 host it's
64-bit, so header decls using 'unsigned long' (e.g. virtual perform(unsigned
long, TGraphics*)) no longer matched the .cpp definitions (which use u32) nor the
base pure virtual perform(u32,...) -> derived classes stayed abstract ('invalid
new-expression of abstract class type' for TViewport/TLightMap/TMapEvent/SunMgr/
... and 'no declaration matches'). Replace with u32 to restore the GC semantics.

+++ include/GC2D/CardSave.hpp
@@ -24,9 +24,9 @@ public:
 	void load(JSUMemoryInputStream& stream);
 	void initData(TMarioGamePad*);
 	void init(int);
-	void perform(unsigned long, JDrama::TGraphics*);
+	void perform(u32, JDrama::TGraphics*);
 	void makeBuffer(J2DTextBox*, int);
-	void setMessage(J2DTextBox*, long, unsigned long);
+	void setMessage(J2DTextBox*, long, u32);
 	s8 waitForStop(TEProgress);
 	void endWaitForChoice();
 	s8 waitForChoice(TEProgress, TEProgress, signed char);
@@ -40,7 +40,7 @@ public:
 	s8 waitForSelect3(TEProgress, TEProgress, TEProgress);
 	s8 waitForAnyKeyBM(TEProgress);
 	void selectBookmarks(TEProgress, TEProgress, TEProgress, TEProgress);
-	void changePattern(J2DPicture*, short, unsigned long);
+	void changePattern(J2DPicture*, short, u32);
 	void execMovement_();
 	u8 getNextState();
 	void execIssueGX_(JDrama::TGraphics*);
