+++ include/JSystem/JDrama/JDRViewport.hpp
@@ -12,7 +12,7 @@ public:
 
 	virtual ~TViewport() { }
 	virtual void load(JSUMemoryInputStream&);
-	virtual void perform(unsigned long, TGraphics*);
+	virtual void perform(u32, TGraphics*);
 
 public:
 	/* 0x10 */ TRect unk10;
